rename livestream + livestreamsetup pages
